Fix for fault [20031011.002].  What was happening was that during the generation of the XML, the UIST and Michelle CalObs in the OT call the method updateDAConf, which resets the exposure time and coadds.
To overcome this, I have added an attribute whenever something is overridden called override_<attribute>.  The existence and value is checked by updateDAConf and if set to true, DAConf does not alter the existing values.  ONLY the QT has the ability to set this flag - it is not user definable.


git-svn-id: file:///jac_sw/svnroot/scicom/trunk/omp/java/qt@6591 64693ad5-8cdc-4012-857f-4cd9bf50b8c8