Make sure we fix up all calibrations, not just non MSBs.
Calibrations are handled differently by JCMTCAL and UKIRTCAL. The QT
handles the two differently for this reason. Either way we want to make
sure it gets fixed up prior to addition or making it persist.

The model used by the list never changes, so we can just use that class
wide rather than ask for it each time. The list is static, so the model
can be also.