Add highlighting of OR folder MSBs.
Fault 20070515.007 requests that when an MSB from an OR
folder is observed, the other MSBs in that folder are
highlighted.  However that is not possible with the
current database schema as discussed in the fault.

This commit introduces a pale green highlight for
all OR folder MSBs in the same project.  This should
alert the operator to the presence of OR folders
so that they can attempt to avoid observing excess
MSBs from one folder.