Fetch MSB column information before making GUI.
This allows us to have the "Splash" window showing its progress
bar animation while we fetch the information from the database.
This requires the Swing event thread to be free, so we have
to do it before starting to construct the GUI in the Swing
thread.

Also update the "Splash" window to use Swing components, increased
the size to ensure the text still fits, and made it use the
Swing event thread in its GUI activities.