JCMT: add JLS restricted widget configuration
Added a new widget configuration file for JLS which is a
copy of the normal JCMT widget configuration but
with JLS as the only available country and semester.
A new command line option -jls (which must be given after
the "jcmt" parameter) activates this configuration.

Requested by JTD, for use with Extended Observing.