UKIRT Translator: read output paths from QT config
The (standalone) UKIRT Translator now reads the QT configuration
file.  This defines the output paths EXEC_PATH and CONF_PATH, so
it is no longer necessary to specify these explicitly in the
launch script.  Additionally the translator no longer defaults
to the user's home directory if the output path option (-o) is
not specified.  The result should be that not specifying an
output path causes the output to be written to the correct
directories.