UKIRT Translator: add prepare for translation step
When the QT fetches an MSB from the database (for execution),
the OMP actually edits the observations in the MSB to include
additional information.  The lack of this information can cause a
problem for the stand-alone UKIRT translator because it needs to
be able to handle files which have been saved directly by the OT,
and which have therefore not been through the OMP system.  Therefore
an additional "prepareForTranslation" method has been added
to the stand-alone UKIRT translator.  This method should include
any steps which the OMP would normally perform that are needed
for translation.

Currently the method adds a "project" attribute to all observations
based on the "projectID" of the science program.  This is because
only "projectID" is present in "OT XML" but the translator relies
on "project" attributes.

For information about what else the OMP would normally do, please see
the OMP::MSB::addFITStoObs method in the OMP Perl code repository.