Use Swing version of SwingWorker
This allows the local version of SwingWorker to be removed.  The
Swing version does not allow the worker to be executed more than
once, so for the usage in QtFrame it was necessary to extract
the code which prepares the worker into a method rather than
having the worker be a member of the class which can be executed
multiple times.