Add copyright statement to XmlUtils.java
The original version of this file appears to have come from the
"XML and Java Tutorial, Part 1" tutorial on http://developerlife.com/.
The copyright statement from the version of the file available from
the tutorial has been added to the file in this repository.

According to http://developerlife.com/tutorials/, "All the source code
provided on this site is distributed under the Apache Open Source
License v2.0."  Therefore the license notice for this license
has also been added to the file, along with a notice stating that
the JAC modified the file, as required by the license.