Make country selection a checkbox group for UKIRT
Changed the widget configuration for UKIRT so that the country selection
is a JCheckBoxGroup instead of a JRadioButtonGroup.  Making it appear in
the same place as the JRadioButtonGroup required changes to the
WidgetPanel.parseConfig method.  The code was first refactored slightly
to take advantage of the commonality between the blocks which created
JCheckBoxGroups and JRadioButtonGroup/JTextFieldGroups.  The member
variable instrumentPanel is now only assigned if we find a group which
is specifically called "Instruments".

Also updated the ButtonPanel methods and query XML generation method
to take account of the new "Any Country" checkbox.

Change requested by Watson:

    Date: Wed, 26 Nov 2014 13:59:34 -1000
    From: Watson Varricatt <w.varricatt@ukirt.hawaii.edu>

    Just had a discussion with Tom about this.  Can we please go ahead with
    check boxes for multi-country selection on the ukirtqt if that is
    feasible to implement.