Move dummy SpProg creation to ExecuteJCMT
When executing deferred observations they need to be placed inside
a SpProg.  However the code to do so was duplicated between
DeferredProgramList and ProgramTree.  This commit moves the
code in question to the ExecuteJCMT class and has it trigger
automatically just before the XML is written out if a deferred
observation is being executed.