Make Execute a subclass of SwingWorker
This is intended to address 20160701.002 by allowing us to place
Swing operations in the "done" method, ensuring that they are
performed in the appropriate thread.

Both the JCMT and UKIRT doInBackground methods return a boolean
status, eliminating the need for success and fail files to be
written.  (Previously the JCMT run method returned a boolean
whereas the UKIRT equivalent could not as it was implementing the
Runnable interface.)  Raising specific exceptions would be
more useful in some cases, such as when there is information
text such as command output which could be displayed.