More consistent loading of icons
Now use ClassLoader.getSystemResource to get the URL necessary
to load icons from the JAR files.  This avoids the need to store
the full path to the image files.

The cut, copy and paste menu item icons previously had the wrong
path and did not appear.

The observation icon used a non-existant IMAG_PATH setting and
so was not being found.  It is now loaded directly from the OT
JAR file.

The selected form of the trash can icon is never used.