Use exception for error in Execute.doInBackground
Execute is now a SwingWorker<Void, Void> rather than a
SwingWorker<Boolean, Void> where instead of returning a boolean status,
it raises exceptions of a new type Execute.SendToQueueExeception in the
event of failure.  This allows informative error messages to more easily
be retrieved for display in the error dialog box (via the exception's
message string).