Use reading threads for Execute.executeCommand
It looks like there's no easy way to do this in Java: when executing a
command, we shouldn't block on reading either standard output or standard
error while the other might be filling up its buffer in order to avoid
the possiblity of deadlock -- instead both should be read together.

This commit adds a stream reading thread and runs two instances of it:
one for each of standard output and standard error of the command being
executed.

This also corrects the handling of the byte[] (now char[]) used to
store standard output -- previously this array was repeatedly
overwritten by multiple reads from the stream and then the array
was returned in whatever state the final read left it.  Now actually
return the text as accumulated in a StringBuffer.