Tidy LogoPanel class
The "buildUI" method has been merged with the constructor.  (There is
no longer any reason for these to be separate.)

The "start" and "startAnimation" and "stop" and "stopAnimation" methods
have been combined.  (Again there didn't appear to be any reason for
there to be two of each.)

Variables are now declared private, and the frameNumber is no longer
static.  The "delay" variable is now a local variable within the
constructor.  (It was never used as an instance variable outside
the old "buildUI" method.)  The "frozen" attribute has been removed
since it looks like it is never being set.