Adjust usage of SwingWorker in QtFrame
The "performSendToStagingArea" method has been adjusted so that
Swing actions are performed either before invoking the SwingWorker,
or in its "done" method.  (Swing actions could otherwise be performed
on the wrong thread.)  The SwingWorker "doInBackground" method now
returns the SpItem directly rather than just returning a Boolean
status.  Exceptions are allowed to propagate to the "done" method
(where they are re-raised at the point of calling "get") since this
allows a message dialog to be shown from the correct thread.