Pass SpItem directly to OmpOM.addNewTree
Now instead of calling "setSpItem" and then calling "addNewTree"
with an integer object, the value of which was ignored, just call
"addNewTree" with the SpItem in question.  This method now also
invokes "updateNotes" (which is now private) automatically with the
given SpItem.

This removes the OmpOM.spItem attribute and an unused method
("resetTree") which used to use it.