Add special sorting for coordinate columns
The ra, dec, ha, az, el and airmass columns are strings because they could
potentially be a /-separated list of coordinates, or a special value such
as "CAL" or "TLE".  So try to take the first component of the list and parse
it as a numerical value.  This should allow us to arrange those values which
are numerical into the correct order.  (They were previously put in
alphabetical order.)

TableSorter currently only has access to the results table via the
Swing TableModel.  Unfortunately this means that we currently have to
identify thse columns via the column name.