Improve sending of multiple coordinates by SAMP
Now attempt to send at least the first set of coordinates
where there are different numbers of RA and Dec values.
If there is only one of one of these, use it for all
sets.  Also split the target names, appending a number
if there are multiple results but not the same number
of names.